requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alembic==1.14.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
huggingface-hub==0.27.0
hyperpyyaml==1.2.2
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy-loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
mako==1.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper==20240930
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
primepy==1.3
propcache==0.2.1
protobuf==5.29.1
pyannote-audio==3.3.2
pyannote-core==5.0.0
pyannote-database==5.1.0
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.8.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.12
scikit-learn==1.6.0
scipy==1.13.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
sqlalchemy==2.0.36
sympy==1.13.1
tabulate==0.9.0
tensorboardx==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1+cu121
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.5
torchaudio==2.5.1+cu121
torchmetrics==1.6.0
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
zipp==3.21.0
